Move JSON converters around for the upcoming ones.
