+++ McMerchants/Startup.cs
@@ -1,10 +1,9 @@
 using McMerchants.Extensions.DependencyInjection;
-using McMerchants.Json;
+using McMerchants.Json.Stock;
 using McMerchants.Services;
 using McMerchants.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
